Split Server.Class to 2 modules
