+++ src/Server/Endpoints/Ping.hs
@@ -6,7 +6,7 @@ module Server.Endpoints.Ping where
 
 import           Utils.Logger       (HasLogger(logMsg))
 import           Servant            (type (:>), NoContent(..), JSON, Get)
-import           Server.Class       (NetworkM)
+import           Server.Internal    (NetworkM)
 
 type PingApi = "relayRequestPing" :> Get '[JSON] NoContent
