+++ src/Server/Endpoints/Tx/New.hs
@@ -12,7 +12,7 @@ import           Control.Monad.Catch              (handle)
 import           Servant                          (JSON, (:>), ReqBody, respond, StdMethod(POST), UVerb, Union)
 import           Server.Endpoints.Tx.Class        (HasTxEndpoints(..))
 import           Server.Endpoints.Tx.Internal     (NewTxEndpointResult(..))
-import           Server.Class                     (NetworkM, HasServer(..))
+import           Server.Internal                  (NetworkM, HasServer(..))
 import           Server.Tx                        (mkBalanceTx)
 import           Utils.Logger                     (HasLogger(..), (.<))
 import           Utils.Servant                    (respondWithStatus)
