+++ src/Server/Endpoints/Tx/Submit.hs
@@ -21,7 +21,8 @@ import           Data.Sequence                    (Seq(..), (|>))
 import           IO.Wallet                        (HasWallet(..))
 import           Servant                          (NoContent(..), JSON, (:>), ReqBody, respond, StdMethod(POST), UVerb, Union)
 import           Server.Endpoints.Tx.Class        (HasTxEndpoints(..))     
-import           Server.Class                     (getQueueRef, NetworkM, Env(..), HasServer(..), QueueRef, checkForCleanUtxos, QueueElem, Queue)
+import           Server.Internal                  (getQueueRef, NetworkM, Env(..), HasServer(..), QueueRef, checkForCleanUtxos,
+                                                   QueueElem, Queue)
 import           Server.Tx                        (mkTx)
 import           Utils.ChainIndex                 (MapUTXO)
 import           Utils.Logger                     (HasLogger(..), (.<), logSmth)
